script: Append to existing CSP list in set_csp_list
