+++ tests/wpt/tests/content-security-policy/meta/meta-append-header.html.headers
@@ -1,0 +1,1 @@
+Content-Security-Policy: script-src 'nonce-abc'
